Switch to LEFT OUTER JOIN for max compatibility
I found that even using LAST_VALUE would break the script in the case of
hash values that are functions that reference other fields, generating a
"this field is not grouped" error. The LEFT OUTER JOIN method is widely
accepted/well tested, and it is the most "standard SQL" type of approach
that will/should work the same way across all platforms. It also performs
about the same as the window functions, as far as I can tell.